+++ launcher-system/src/main/java/com/pactera/business/dao/LaunThemeMapper.java
@@ -109,6 +109,6 @@ public interface LaunThemeMapper extends BaseMapper<LaunThemeAdministration> {
 
 	int cleanClassification(@Param("id") String id);
 
-	List<LaunResolution> resolution(@Param("layoutId") Integer layoutId);
+	List<LaunResolution> resolution(@Param("layoutId") Long layoutId);
 
 }
